requirements.txt
Pillow
torch
tensorboard
torchvision
